Add Animations on General Settings Screen, Remove Wine Settings and Display Settings and Unify Into Debug Settings, Re-Add Debug Settings for Box64 on Debug Settings
