+++ app/src/main/java/com/micewine/emu/activities/GeneralSettingsActivity.kt
@@ -15,19 +15,17 @@ import com.micewine.emu.R
 import com.micewine.emu.activities.MainActivity.Companion.setSharedVars
 import com.micewine.emu.databinding.ActivityGeneralSettingsBinding
 import com.micewine.emu.fragments.Box64SettingsFragment
-import com.micewine.emu.fragments.DisplaySettingsFragment
+import com.micewine.emu.fragments.DebugSettingsFragment
 import com.micewine.emu.fragments.DriversSettingsFragment
 import com.micewine.emu.fragments.EnvVarsSettingsFragment
 import com.micewine.emu.fragments.GeneralSettingsFragment
 import com.micewine.emu.fragments.SoundSettingsFragment
-import com.micewine.emu.fragments.WineSettingsFragment
 
 class GeneralSettingsActivity : AppCompatActivity() {
     private var binding: ActivityGeneralSettingsBinding? = null
     private var backButton: ImageButton? = null
     private val box64SettingsFragment = Box64SettingsFragment()
-    private val wineSettingsFragment = WineSettingsFragment()
-    private val displaySettingsFragment = DisplaySettingsFragment()
+    private val debugSettingsFragment = DebugSettingsFragment()
     private val driversSettingsFragment = DriversSettingsFragment()
     private val environmentVariablesSettings = EnvVarsSettingsFragment()
     private val soundSettingsFragment = SoundSettingsFragment()
@@ -43,16 +41,10 @@ class GeneralSettingsActivity : AppCompatActivity() {
                         fragmentLoader(box64SettingsFragment, false)
                     }
 
-                    getString(R.string.wine_settings_title) -> {
-                        generalSettingsToolbar?.title = getString(R.string.wine_settings_title)
+                    getString(R.string.debug_settings_title) -> {
+                        generalSettingsToolbar?.title = getString(R.string.debug_settings_title)
 
-                        fragmentLoader(wineSettingsFragment, false)
-                    }
-
-                    getString(R.string.display_settings_title) -> {
-                        generalSettingsToolbar?.title = getString(R.string.display_settings_title)
-
-                        fragmentLoader(displaySettingsFragment, false)
+                        fragmentLoader(debugSettingsFragment, false)
                     }
 
                     getString(R.string.driver_settings_title) -> {
@@ -120,6 +112,12 @@ class GeneralSettingsActivity : AppCompatActivity() {
 
     private fun fragmentLoader(fragment: Fragment, appInit: Boolean) {
         supportFragmentManager.beginTransaction().apply {
+            setCustomAnimations(
+                R.anim.slide_in,
+                R.anim.fade_out,
+                R.anim.fade_in,
+                R.anim.slide_out
+            )
             replace(R.id.settings_content, fragment)
 
             if (!appInit) {
@@ -167,8 +165,6 @@ class GeneralSettingsActivity : AppCompatActivity() {
         const val BOX64_DYNAREC_ALIGNED_ATOMICS_DEFAULT_VALUE = false
         const val BOX64_DYNAREC_NATIVEFLAGS = "BOX64_DYNAREC_NATIVEFLAGS"
         const val BOX64_DYNAREC_NATIVEFLAGS_DEFAULT_VALUE = true
-        const val BOX64_DYNAREC_BLEEDING_EDGE = "BOX64_DYNAREC_BLEEDING_EDGE"
-        const val BOX64_DYNAREC_BLEEDING_EDGE_DEFAULT_VALUE = true
         const val BOX64_DYNAREC_WAIT = "BOX64_DYNAREC_WAIT"
         const val BOX64_DYNAREC_WAIT_DEFAULT_VALUE = true
         const val BOX64_DYNAREC_DIRTY = "BOX64_DYNAREC_DIRTY"
@@ -188,8 +184,6 @@ class GeneralSettingsActivity : AppCompatActivity() {
         const val SELECTED_WINE_PREFIX = "selectedWinePrefix"
         const val SELECTED_TU_DEBUG_PRESET = "selectedTuDebugPreset"
         const val SELECTED_TU_DEBUG_PRESET_DEFAULT_VALUE = "noconform,sysmem"
-        const val SELECTED_DRIVER = "selectedDriver"
-        const val SELECTED_DRIVER_DEFAULT_VALUE = ""
         const val ENABLE_DRI3 = "enableDRI3"
         const val ENABLE_DRI3_DEFAULT_VALUE = true
         const val ENABLE_MANGOHUD = "enableMangoHUD"
