+++ app/src/main/res/anim/fade_out.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<alpha xmlns:android="http://schemas.android.com/apk/res/android"
+    android:duration="@android:integer/config_longAnimTime"
+    android:interpolator="@android:anim/decelerate_interpolator"
+    android:fromAlpha="1"
+    android:toAlpha="0" />
