+++ app/src/main/res/anim/slide_in.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<translate xmlns:android="http://schemas.android.com/apk/res/android"
+    android:duration="@android:integer/config_longAnimTime"
+    android:interpolator="@android:anim/decelerate_interpolator"
+    android:fromXDelta="100%"
+    android:toXDelta="0%" />
